+++ H750VBT6-flight-board-v0/System/States/state_pre_flight.h
@@ -31,7 +31,7 @@ class PreFlightState : public State {
 
 		static constexpr double kLaunchAccelThreshold = 20.0; // m/s**2
 		static constexpr double kLaunchPosZDiffThreshold = 1.0; // m
-		static constexpr double kTransitionResetTimeThreshold = 500; // 500 ms transition timer
+		static constexpr uint32_t kTransitionResetTimeThreshold = 500; // 500 ms transition timer
 
 		uint32_t transitionResetTimer;
 		double minPosZ = 1000000;
